done and jnew

+++ src/components/clients/client.service.ts
@@ -101,36 +101,38 @@ export class ClientService {
                 const today = (new Date(Date.now())).toISOString().split('T')[0];
                 const existingClientUser = (await this.usersService.search({ mobile: existingClientMobile }))[0];
                 try {
-                    await this.telegramService.createClient(existingClientMobile, false, true)
-                    if (toBoolean(setupClientQueryDto.formalities)) {
-                        console.log("Started Formalities")
-                        await this.telegramService.updateUsername(existingClientMobile, '');
-                        await sleep(2000)
-                        await this.telegramService.updatePrivacyforDeletedAccount(existingClientMobile)
-                        await sleep(2000)
-                        await this.telegramService.deleteProfilePhotos(existingClientMobile)
-                        console.log("Formalities finished")
-                        await fetchWithTimeout(`${ppplbot()}&text=Formalities finished`);
-                    } else {
-                        console.log("Formalities skipped")
-                    }
-                    if (archiveOld) {
-                        const availableDate = (new Date(Date.now() + (setupClientQueryDto.days * 24 * 60 * 60 * 1000))).toISOString().split('T')[0]
-                        const bufferClientDto: CreateBufferClientDto | UpdateBufferClientDto = {
-                            mobile: existingClientMobile,
-                            createdDate: today,
-                            updatedDate: today,
-                            availableDate,
-                            session: existingClientUser.session,
-                            tgId: existingClientUser.tgId,
-                            channels: 100
+                    if (existingClientUser) {
+                        await this.telegramService.createClient(existingClientMobile, false, true)
+                        if (toBoolean(setupClientQueryDto.formalities)) {
+                            console.log("Started Formalities")
+                            await this.telegramService.updateUsername(existingClientMobile, '');
+                            await sleep(2000)
+                            await this.telegramService.updatePrivacyforDeletedAccount(existingClientMobile)
+                            await sleep(2000)
+                            await this.telegramService.deleteProfilePhotos(existingClientMobile)
+                            console.log("Formalities finished")
+                            await fetchWithTimeout(`${ppplbot()}&text=Formalities finished`);
+                        } else {
+                            console.log("Formalities skipped")
                         }
-                        const updatedBufferClient = await this.bufferClientService.createOrUpdate(existingClientMobile, bufferClientDto);
-                        await this.archivedClientService.update(existingClient.mobile, existingClient);
-                        console.log("client Archived: ", updatedBufferClient)
-                        await fetchWithTimeout(`${ppplbot()}&text=Client Archived`);
-                    } else {
-                        console.log("Client Archive Skipped")
+                        if (archiveOld) {
+                            const availableDate = (new Date(Date.now() + (setupClientQueryDto.days * 24 * 60 * 60 * 1000))).toISOString().split('T')[0]
+                            const bufferClientDto: CreateBufferClientDto | UpdateBufferClientDto = {
+                                mobile: existingClientMobile,
+                                createdDate: today,
+                                updatedDate: today,
+                                availableDate,
+                                session: existingClientUser.session,
+                                tgId: existingClientUser.tgId,
+                                channels: 100
+                            }
+                            const updatedBufferClient = await this.bufferClientService.createOrUpdate(existingClientMobile, bufferClientDto);
+                            await this.archivedClientService.update(existingClient.mobile, existingClient);
+                            console.log("client Archived: ", updatedBufferClient)
+                            await fetchWithTimeout(`${ppplbot()}&text=Client Archived`);
+                        } else {
+                            console.log("Client Archive Skipped")
+                        }
                     }
                 } catch (error) {
                     console.log("Cannot Archive Old Client");
